Move constants to recentMaps constants
